policy-procedure-permissions: added edit permission on policy and procedures

+++ apps/console/src/components/pages/protected/policies/view/view-policy-page.tsx
@@ -19,7 +19,7 @@ import { useQueryClient } from '@tanstack/react-query'
 import { useNotification } from '@/hooks/useNotification.tsx'
 import { usePolicy } from '@/components/pages/protected/policies/create/hooks/use-policy.tsx'
 import AssociatedObjectsViewAccordion from '@/components/pages/protected/policies/accordion/associated-objects-view-accordion.tsx'
-import { canDelete } from '@/lib/authz/utils'
+import { canDelete, canEdit } from '@/lib/authz/utils'
 import { useAccountRole } from '@/lib/authz/access-api'
 import { useSession } from 'next-auth/react'
 import { ObjectEnum } from '@/lib/authz/enums/object-enum'
@@ -44,6 +44,7 @@ const ViewPolicyPage: React.FC<TViewPolicyPage> = ({ policyId }) => {
   const { data: session } = useSession()
   const { data: permission } = useAccountRole(session, ObjectEnum.POLICY, policyId)
   const deleteAllowed = canDelete(permission?.roles)
+  const editAllowed = canEdit(permission?.roles)
   const { mutateAsync: deletePolicy } = useDeleteInternalPolicy()
   const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false)
   const router = useRouter()
@@ -169,30 +170,36 @@ const ViewPolicyPage: React.FC<TViewPolicyPage> = ({ policyId }) => {
                 </div>
               ) : (
                 <div className="flex gap-2 justify-end">
-                  <Menu
-                    content={
-                      <>
-                        <div className="flex items-center space-x-2 cursor-pointer" onClick={handleEdit}>
-                          <PencilIcon size={16} strokeWidth={2} />
-                          <span>Edit</span>
-                        </div>
-                        {deleteAllowed && (
-                          <>
-                            <div className="flex items-center space-x-2 cursor-pointer" onClick={() => setIsDeleteDialogOpen(true)}>
-                              <Trash2 size={16} strokeWidth={2} />
-                              <span>Delete</span>
+                  {!editAllowed && !deleteAllowed ? (
+                    <></>
+                  ) : (
+                    <Menu
+                      content={
+                        <>
+                          {editAllowed && (
+                            <div className="flex items-center space-x-2 cursor-pointer" onClick={handleEdit}>
+                              <PencilIcon size={16} strokeWidth={2} />
+                              <span>Edit</span>
                             </div>
-                            <ConfirmationDialog
-                              open={isDeleteDialogOpen}
-                              onOpenChange={setIsDeleteDialogOpen}
-                              onConfirm={handleDeletePolicy}
-                              description="This action cannot be undone. This will permanently remove the policy from the organization."
-                            />
-                          </>
-                        )}
-                      </>
-                    }
-                  />
+                          )}
+                          {deleteAllowed && (
+                            <>
+                              <div className="flex items-center space-x-2 cursor-pointer" onClick={() => setIsDeleteDialogOpen(true)}>
+                                <Trash2 size={16} strokeWidth={2} />
+                                <span>Delete</span>
+                              </div>
+                              <ConfirmationDialog
+                                open={isDeleteDialogOpen}
+                                onOpenChange={setIsDeleteDialogOpen}
+                                onConfirm={handleDeletePolicy}
+                                description="This action cannot be undone. This will permanently remove the policy from the organization."
+                              />
+                            </>
+                          )}
+                        </>
+                      }
+                    />
+                  )}
                 </div>
               )}
               <AuthorityCard form={form} approver={policy.approver} delegate={policy.delegate} isEditing={isEditing} />
