+++ apps/console/src/components/pages/protected/procedures/view/view-procedure-page.tsx
@@ -7,7 +7,7 @@ import { Form } from '@repo/ui/form'
 import DetailsField from '@/components/pages/protected/procedures/view/fields/details-field.tsx'
 import TitleField from '@/components/pages/protected/procedures/view/fields/title-field.tsx'
 import { Button } from '@repo/ui/button'
-import { CirclePlus, PencilIcon, SaveIcon, XIcon } from 'lucide-react'
+import { PencilIcon, SaveIcon, XIcon } from 'lucide-react'
 import AuthorityCard from '@/components/pages/protected/procedures/view/cards/authority-card.tsx'
 import PropertiesCard from '@/components/pages/protected/procedures/view/cards/properties-card.tsx'
 import HistoricalCard from '@/components/pages/protected/procedures/view/cards/historical-card.tsx'
@@ -26,7 +26,7 @@ import { ConfirmationDialog } from '@repo/ui/confirmation-dialog'
 import { useSession } from 'next-auth/react'
 import { useAccountRole } from '@/lib/authz/access-api'
 import { ObjectEnum } from '@/lib/authz/enums/object-enum'
-import { canDelete } from '@/lib/authz/utils'
+import { canDelete, canEdit } from '@/lib/authz/utils'
 import { useDeleteProcedure } from '@/lib/graphql-hooks/procedures'
 import Menu from '@/components/shared/menu/menu.tsx'
 
@@ -48,6 +48,7 @@ const ViewProcedurePage: React.FC<TViewProcedurePage> = ({ procedureId }) => {
   const { data: session } = useSession()
   const { data: permission } = useAccountRole(session, ObjectEnum.PROCEDURE, procedureId)
   const deleteAllowed = canDelete(permission?.roles)
+  const editAllowed = canEdit(permission?.roles)
   const { mutateAsync: deleteProcedure } = useDeleteProcedure()
   const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false)
 
@@ -173,30 +174,36 @@ const ViewProcedurePage: React.FC<TViewProcedurePage> = ({ procedureId }) => {
                   </div>
                 ) : (
                   <div className="flex gap-2 justify-end">
-                    <Menu
-                      content={
-                        <>
-                          <div className="flex items-center space-x-2 cursor-pointer" onClick={handleEdit}>
-                            <PencilIcon size={16} strokeWidth={2} />
-                            <span>Edit</span>
-                          </div>
-                          {deleteAllowed && (
-                            <>
-                              <div className="flex items-center space-x-2 cursor-pointer" onClick={() => setIsDeleteDialogOpen(true)}>
-                                <Trash2 size={16} strokeWidth={2} />
-                                <span>Delete</span>
+                    {!editAllowed && !deleteAllowed ? (
+                      <></>
+                    ) : (
+                      <Menu
+                        content={
+                          <>
+                            {editAllowed && (
+                              <div className="flex items-center space-x-2 cursor-pointer" onClick={handleEdit}>
+                                <PencilIcon size={16} strokeWidth={2} />
+                                <span>Edit</span>
                               </div>
-                              <ConfirmationDialog
-                                open={isDeleteDialogOpen}
-                                onOpenChange={setIsDeleteDialogOpen}
-                                onConfirm={handleDeleteProcedure}
-                                description="This action cannot be undone. This will permanently remove the procedure from the organization."
-                              />
-                            </>
-                          )}
-                        </>
-                      }
-                    />
+                            )}
+                            {deleteAllowed && (
+                              <>
+                                <div className="flex items-center space-x-2 cursor-pointer" onClick={() => setIsDeleteDialogOpen(true)}>
+                                  <Trash2 size={16} strokeWidth={2} />
+                                  <span>Delete</span>
+                                </div>
+                                <ConfirmationDialog
+                                  open={isDeleteDialogOpen}
+                                  onOpenChange={setIsDeleteDialogOpen}
+                                  onConfirm={handleDeleteProcedure}
+                                  description="This action cannot be undone. This will permanently remove the procedure from the organization."
+                                />
+                              </>
+                            )}
+                          </>
+                        }
+                      />
+                    )}
                   </div>
                 )}
               </div>
